Add files via upload
Python Email slicer is basically made with RegEx library and it does slices the email address in to parts the first is the name and second is the domain name. This function takes a file name or location of the file which contains any numbers of email addresses after the code is run. A file will be  made which contains the domain name and name of the email address.

Syed Muhammad Rimmel Asghar